Create 8-bit-game-using-pygame
Pygame is a python library that can be used specifically to design and build games. Pygame supports only 2d games that are built using different sprites. Pygame is not particularly best for designing games as it is very complex to use doesn’t have a proper GUI like unity but it definitely builds logic for further complex projects.
We’ll be creating a simple game with the following rules:-
 

The player can only move vertically. 
 
Other than player block there will be two other blocks. 
 
One of them will be the enemy block and one of them will be score block. 
 
If the player collides with the enemy block then the game over screen pops up, if the player collides with the score block the score is incremented and it is compulsory to collect all score blocks.