Create tic-tac-toe-gui-in-python-using-pygame
This article will guide you and give you a basic idea of designing a game Tic Tac Toe using pygame library of Python. Pygame is a cross-platform set of Python modules designed for writing video games. It includes computer graphics and sound libraries designed to be used with the Python programming language. Let’s break the task in five parts:

Importing the required libraries and setting up the required global variables.
Designing the game display function, that will set a platform for other components to be displayed on the screen.
Main algorithm of win and draw
Getting the user input and displaying the “X” or “O” at the proper position where the user has clicked his mouse.
Running an infinite loop, and including the defined methods in it.